+++ src/utils.h
@@ -239,7 +239,7 @@ size_varsize(size_t val) {
 }
 
 static inline bool
-read_u8(volatile uint8_t **buf, uint8_t *len, uint8_t *u8) {
+read_u8(volatile uint8_t **buf, uint16_t *len, uint8_t *u8) {
   if (*len < 1)
     return false;
 
@@ -251,7 +251,7 @@ read_u8(volatile uint8_t **buf, uint8_t *len, uint8_t *u8) {
 }
 
 static inline bool
-read_u16(volatile uint8_t **buf, uint8_t *len, uint16_t *u16, bool be) {
+read_u16(volatile uint8_t **buf, uint16_t *len, uint16_t *u16, bool be) {
   if (*len < 2)
     return false;
 
@@ -270,7 +270,7 @@ read_u16(volatile uint8_t **buf, uint8_t *len, uint16_t *u16, bool be) {
 }
 
 static inline bool
-read_u32(volatile uint8_t **buf, uint8_t *len, uint32_t *u32, bool be) {
+read_u32(volatile uint8_t **buf, uint16_t *len, uint32_t *u32, bool be) {
   if (*len < 4)
     return false;
 
@@ -291,7 +291,7 @@ read_u32(volatile uint8_t **buf, uint8_t *len, uint32_t *u32, bool be) {
 }
 
 static inline bool
-read_varint(volatile uint8_t **buf, uint8_t *len, hns_varint_t *varint) {
+read_varint(volatile uint8_t **buf, uint16_t *len, hns_varint_t *varint) {
   if (*len < 1)
     return false;
 
@@ -350,7 +350,7 @@ read_varint(volatile uint8_t **buf, uint8_t *len, hns_varint_t *varint) {
 }
 
 static inline bool
-peek_varint(volatile uint8_t **buf, uint8_t *len, hns_varint_t *varint) {
+peek_varint(volatile uint8_t **buf, uint16_t *len, hns_varint_t *varint) {
   if (!read_varint(buf, len, varint))
     return false;
 
@@ -363,7 +363,7 @@ peek_varint(volatile uint8_t **buf, uint8_t *len, hns_varint_t *varint) {
 }
 
 static inline bool
-read_varsize(volatile uint8_t **buf, uint8_t *len, size_t *val) {
+read_varsize(volatile uint8_t **buf, uint16_t *len, size_t *val) {
   hns_varint_t v;
 
   if (!read_varint(buf, len, &v))
@@ -375,7 +375,7 @@ read_varsize(volatile uint8_t **buf, uint8_t *len, size_t *val) {
 }
 
 static inline bool
-read_bytes(volatile uint8_t **buf, uint8_t *len, volatile uint8_t *out, size_t sz) {
+read_bytes(volatile uint8_t **buf, uint16_t *len, volatile uint8_t *out, size_t sz) {
   if (*len < sz)
     return false;
 
@@ -390,7 +390,7 @@ read_bytes(volatile uint8_t **buf, uint8_t *len, volatile uint8_t *out, size_t s
 static inline bool
 read_varbytes(
   volatile uint8_t **buf,
-  uint8_t *len,
+  uint16_t *len,
   uint8_t *out,
   size_t out_sz,
   size_t *out_len
@@ -412,6 +412,7 @@ read_varbytes(
   if (!read_bytes(buf, len, out, sz)) {
     *buf -= offset;
     *len += offset;
+
     return false;
   }
 
@@ -423,7 +424,7 @@ read_varbytes(
 static inline bool
 read_bip44_path(
   volatile uint8_t **buf,
-  uint8_t *len,
+  uint16_t *len,
   uint8_t *depth,
   uint32_t *path,
   uint8_t *info
